modifies hypothesis widening view
