+++ lib/pages/views/issue_page_views/widening_hypotheses_view.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:guide_solve/bloc/issue/issue_bloc.dart';
 import 'package:guide_solve/components/issue_solving_widgets/process_status_bar.dart';
-import 'package:guide_solve/components/issue_solving_widgets/vote_button_hypotheses_widen.dart';
 import 'package:guide_solve/models/hypothesis.dart';
 import 'package:guide_solve/models/issue.dart';
 
@@ -132,77 +131,169 @@ Widget _hypothesisList(
     TextEditingController textController,
     FocusNode focusNode) {
   return Expanded(
-      child: StreamBuilder<List<Hypothesis>>(
-    stream: issueBloc.hypothesesStream,
-    builder: (context, hypothesesSnapshot) {
-      if (hypothesesSnapshot.hasError) {
-        return const Center(
-          child: Text('Error loading hypotheses'),
-        );
-      }
-      if (!hypothesesSnapshot.hasData) {
-        return const Center(
-          child: Text('Submit a root issue theory.'),
-        );
-      }
-      final hypotheses = hypothesesSnapshot.data!;
+    child: StreamBuilder<List<Hypothesis>>(
+      stream: issueBloc.hypothesesStream,
+      builder: (context, hypothesesSnapshot) {
+        if (hypothesesSnapshot.hasError) {
+          return const Center(
+            child: Text('Error loading hypotheses'),
+          );
+        }
+        if (!hypothesesSnapshot.hasData) {
+          return const Center(
+            child: Text('Submit a root issue theory.'),
+          );
+        }
+        final hypotheses = hypothesesSnapshot.data!;
 
-      // Calculate rank for each hypothesis using
-      // Perspective and update rank value
-      for (final hypothesis in hypotheses) {
-        final perspective = hypothesis.perspective(
-          currentUserId,
-          issueBloc.focusedIssue!.invitedUserIds!,
+        // Calculate rank for each hypothesis using
+        // Perspective and update rank value
+        for (final hypothesis in hypotheses) {
+          final perspective = hypothesis.perspective(
+            currentUserId,
+            issueBloc.focusedIssue!.invitedUserIds!,
+          );
+          hypothesis.rank = perspective.calculateConsensusRank();
+        }
+        // Sort hypotheses based on rank in
+        // descending order (higher rank first)
+        hypotheses.sort(
+          (a, b) => b.rank.compareTo(a.rank),
         );
-        hypothesis.rank = perspective.calculateRank(
-            (context.read<IssueBloc>().state as IssueProcessState).stage);
-      }
-      // Sort hypotheses based on rank in
-      // descending order (higher rank first)
-      hypotheses.sort(
-        (a, b) => b.rank.compareTo(a.rank),
-      );
 
-      return Align(
-        child: ConstrainedBox(
-          constraints: const BoxConstraints(maxWidth: 1000),
-          child: ListView.builder(
-            itemCount: hypotheses.length,
-            itemBuilder: (context, index) {
-              final hypothesis = hypotheses[index];
-              final currentUserVote = hypothesis
-                  .perspective(
-                      currentUserId, issueBloc.focusedIssue!.invitedUserIds!)
-                  .getCurrentUserVote();
-              final everyoneElseAgrees = hypothesis
-                  .perspective(
-                      currentUserId, issueBloc.focusedIssue!.invitedUserIds!)
-                  .allOtherStakeholdersAgree();
-              return Padding(
-                padding: const EdgeInsets.symmetric(vertical: AppSpacing.xxs),
-                child: ShadCard(
-                  title: Text(
-                    hypothesis.desc,
-                    style: UITextStyle.subtitle1,
+        return Align(
+          child: ConstrainedBox(
+            constraints: const BoxConstraints(maxWidth: 1000),
+            child: ListView.builder(
+              itemCount: hypotheses.length,
+              itemBuilder: (context, index) {
+                final hypothesis = hypotheses[index];
+                final currentUserVote = hypothesis
+                    .perspective(
+                        currentUserId, issueBloc.focusedIssue!.invitedUserIds!)
+                    .getCurrentUserVote();
+                final everyoneElseAgrees = hypothesis
+                    .perspective(
+                        currentUserId, issueBloc.focusedIssue!.invitedUserIds!)
+                    .allOtherStakeholdersAgree();
+                return Padding(
+                  padding: const EdgeInsets.symmetric(vertical: AppSpacing.xxs),
+                  child: ShadCard(
+                    title: Tappable(
+                      child: Text(
+                        hypothesis.desc,
+                        style: UITextStyle.subtitle1,
+                      ),
+                      onLongPress: () {
+                        textController.text = hypothesis.desc;
+                        focusNode.requestFocus();
+                      },
+                    ),
+                    backgroundColor: currentUserVote == HypothesisVote.spinoff
+                        ? AppColors.conflictLight
+                        : everyoneElseAgrees
+                            ? AppColors.consensus
+                            : AppColors.public,
+                    trailing: WidenHypothesesSegmentButton(
+                      hypothesis: hypothesis,
+                      currentUserId: currentUserId,
+                      invitedUserIds: issueBloc.focusedIssue!.invitedUserIds!,
+                    ),
                   ),
-                  backgroundColor: currentUserVote == HypothesisVote.spinoff
-                      ? AppColors.conflictLight
-                      : everyoneElseAgrees
-                          ? AppColors.consensus
-                          : AppColors.public,
-                  trailing: WidenHypothesesSegmentButton(
-                    hypothesis: hypothesis,
-                    currentUserId: currentUserId,
-                    invitedUserIds: issueBloc.focusedIssue!.invitedUserIds!,
-                    textController: textController,
-                    focusNode: focusNode,
-                  ),
-                ),
-              );
-            },
+                );
+              },
+            ),
+          ),
+        );
+      },
+    ),
+  );
+}
+
+class WidenHypothesesSegmentButton extends StatefulWidget {
+  const WidenHypothesesSegmentButton({
+    required this.hypothesis,
+    required this.currentUserId,
+    required this.invitedUserIds,
+    super.key,
+  });
+
+  final Hypothesis hypothesis;
+  final String currentUserId;
+  final List<String> invitedUserIds;
+
+  @override
+  State<WidenHypothesesSegmentButton> createState() =>
+      _WidenHypothesesSegmentButtonState();
+}
+
+class _WidenHypothesesSegmentButtonState
+    extends State<WidenHypothesesSegmentButton> {
+  HypothesisVote? currentUserVote;
+
+  @override
+  void initState() {
+    super.initState();
+    // Initialize the current vote
+    currentUserVote = widget.hypothesis.votes[widget.currentUserId];
+  }
+
+  void _handleVote(HypothesisVote value) {
+    context.read<IssueBloc>().add(
+          HypothesisVoteSubmitted(
+            voteValue: value,
+            hypothesisId: widget.hypothesis.hypothesisId!,
+          ),
+        );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final hypothesis = widget.hypothesis;
+
+    // Use the latest vote from the hypothesis instead of local state
+    final currentUserVote = hypothesis.votes[widget.currentUserId];
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        SegmentedButton<HypothesisVote>(
+          segments: [
+            ButtonSegment<HypothesisVote>(
+                value: HypothesisVote.disagree,
+                label: const Text('Disagree'),
+                tooltip: 'Disagree with this hypothesis.'),
+            ButtonSegment<HypothesisVote>(
+                value: HypothesisVote.agree,
+                label: const Text('Agree'),
+                tooltip:
+                    'Agree that this hypothesis could be part of the issue.'),
+          ],
+          selected: currentUserVote != null ? {currentUserVote} : {},
+          multiSelectionEnabled: false,
+          emptySelectionAllowed: true,
+          showSelectedIcon: false,
+          onSelectionChanged: (Set<HypothesisVote> newSelection) {
+            if (newSelection.isNotEmpty) {
+              final selectedVote = newSelection.first;
+              _handleVote(selectedVote);
+            }
+          },
+          style: SegmentedButton.styleFrom(
+            padding: EdgeInsets.symmetric(horizontal: 2, vertical: 2),
+            maximumSize: const Size(40, 20),
+            textStyle: const TextStyle(fontSize: 11),
+            backgroundColor:
+                currentUserVote == null ? AppColors.private : AppColors.public,
+            selectedBackgroundColor: currentUserVote == HypothesisVote.agree
+                ? AppColors.consensus
+                : currentUserVote == HypothesisVote.root
+                    ? AppColors.consensus
+                    : currentUserVote == HypothesisVote.disagree
+                        ? AppColors.conflictLight
+                        : AppColors.private,
           ),
         ),
-      );
-    },
-  ));
+      ],
+    );
+  }
 }
